#90 Refactor: TokenController 생성
