+++ src/main/java/com/drinkeg/drinkeg/controller/TokenController.java
@@ -1,0 +1,23 @@
+package com.drinkeg.drinkeg.controller;
+
+import com.drinkeg.drinkeg.apipayLoad.ApiResponse;
+import com.drinkeg.drinkeg.service.loginService.TokenService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+@RestController
+@RequiredArgsConstructor
+public class TokenController {
+
+    private final TokenService tokenService;
+    @PostMapping("/reissue")
+    public ApiResponse<?> reissue(HttpServletRequest request, HttpServletResponse response) {
+
+        tokenService.reissueRefreshToken(request, response);
+        return ApiResponse.onSuccess("토큰 재발급 성공");
+    }
+}
